feat: delete logout from settings
